[ADD] base: allow nulls last ordering
Before this commit it was not possible to order a model by nulls last. In some cases this can be very handy though.
For example: if you want to order a score asc/desc but you want nulls to end up at the end of your filtering. Without 'nulls last' you would get:
name | score
-------------------+----------------------
abc |
def | 6.76
xxx |  5.8

With 'nulls last' you would get:
name | score
-------------------+----------------------
def | 6.76
xxx |  5.8
abc |

After this commit you will be able to also order on nulls last to sort your data in a more flexible way.
Note: it is not planned to expose these technical options in the web clients it's UI!
This is only added for developers so they can use this in custom modules